+++ src/main/java/lumynous/safechat/util/Fmt.java
@@ -1,11 +1,11 @@
-package luminous.safechat.util;
+package lumynous.safechat.util;
 
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
 
 public class Fmt {
-    public static String format(@NotNull String fmt, @NotNull Map<@NotNull String, Object> args) {
+    public static String format(@NotNull String fmt, @NotNull Map<@NotNull String, ?> args) {
         String result = fmt;
         for (var arg : args.entrySet()) {
             if (arg.getValue() == null) continue;