+++ CHANGELOG.md
@@ -1,14 +1,25 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## v1.1.0
-Released on: 2022-08-16
+## [2.0.0] - 2023-05-01
+Released on:
+### Added
+- `reload-config` command to reload the configurations.
+### Fixed
+- Missing spaces in direct messages.
+- Colors used in the format are formatted when the configuration loaded now.
+### Removed
+- Test command.
+
+## [1.1.0] - 2022-08-16
 ### Added
 - Direct message protection.
 - Message format configuration.
 
-## v1.0.0
-Released on: 2022-08-04
+## [1.0.0] - 2022-08-04
 ### Added
 - Message re-sending.
 - Test command.
+
+[1.1.0]: https://github.com/lumynou5/Minecraft-Plugin-Safe-Chat/releases/tag/v1.1.0
+[1.0.0]: https://github.com/lumynou5/Minecraft-Plugin-Safe-Chat/releases/tag/v1.0.0